+++ test/test_pkg_test.py
@@ -5,7 +5,7 @@ import subprocess as sp
 import pytest
 
 from helpers import Recipes, ensure_missing
-from bioconda_utils import pkg_test
+from bioconda_utils import docker_utils, pkg_test
 from bioconda_utils import utils
 from bioconda_utils import build
 
@@ -50,7 +50,7 @@ one:
 
 
 # Skip mulled_test on default since we already run pkg_test.test_package for every test case.
-def _build_pkg(recipe, mulled_test=False):
+def _build_pkg(recipe, mulled_test=False, docker_builder=None):
     r = Recipes(recipe, from_string=True)
     r.write_recipes()
     recipe = r.recipe_dirs['one']
@@ -61,6 +61,7 @@ def _build_pkg(recipe, mulled_test=False):
         recipe=r.recipe_dirs['one'],
         pkg_paths=built_packages,
         mulled_test=mulled_test,
+        docker_builder=docker_builder,
     )
     return built_packages
 
@@ -94,38 +95,3 @@ def test_pkg_test_custom_base_image():
     built_packages = _build_pkg(RECIPE_CUSTOM_BASE)
     for pkg in built_packages:
         pkg_test.test_package(pkg, base_image='debian:latest')
-
-
-@pytest.mark.skipif(SKIP_OSX, reason="skipping on osx")
-def test_pkg_test_conda_image():
-    """
-    Check mulled-build test image has conda/mamba not older than outside build.
-    """
-    # Require versions at least at high as those used by bioconda-utils itself.
-    recipe = dedent(f"""
-        one:
-          meta.yaml: |
-            package:
-              name: test_pkg_test_conda_image
-              version: 0.1
-            requirements:
-              run:
-                - python
-                - setuptools
-            test:
-              commands:
-                - |
-                  python -c '
-                  import sys, os
-                  from pathlib import Path
-                  from pkg_resources import parse_version as v
-                  assert v("{conda_version}") <= v(Path(os.environ["CONDA_PREFIX"], "conda-version").read_text())
-                  assert v("{mamba_version}") <= v(Path(os.environ["CONDA_PREFIX"], "mamba-version").read_text())
-                  '
-          post-link.sh: |
-            conda --version | sed -n 's/^conda //p' > "${{PREFIX}}/conda-version"
-            mamba --version | sed -n 's/^mamba //p' > "${{PREFIX}}/mamba-version"
-    """)  # noqa: E501: line too long
-    built_packages = _build_pkg(recipe)
-    for pkg in built_packages:
-        pkg_test.test_package(pkg)